pyproject.toml
[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
# when upgrading this, make sure to also update Dockerfile base image
python = "^3.12"
python-escpos = "^3.1"
pytz = "^2024.1"
bottle = '^0.12'

[tool.poetry.group.dev.dependencies]
black = { version = '^24.4' }
isort = '^5.13'
Werkzeug = "^3.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  \.git
  | \.mypy_cache
  | build
  | dist
)
'''

[tool.isort]
src_paths = 'starship'
profile = 'black'
skip_gitignore = false

[tool.pylint.MASTER]
load-plugins = [
  'pylint.extensions.docparams',
]

[tool.pylint.'MESSAGES CONTROL']
disable = [
  'fixme',
  'line-too-long',
  'invalid-name',
  'too-few-public-methods',
  'too-many-arguments',
  'too-many-instance-attributes',
  'duplicate-code',
  'missing-docstring',
  'broad-exception-caught',
]

[tool.pylint.DESIGN]
max-parents = 25
ignored-argument-names = '_.*|arg|args|kwargs'

[tool.pylint.REPORTS]
output-format = 'parseable'
reports = 'no'

[tool.pylint.FORMAT]
expected-line-ending-format = 'LF'
max-line-length = 88

[tool.pylint.MISCELLANEOUS]
notes = 'FIXME,XXX'
